requirements.txt
# external requirements
flask
flask-cors
cianparser
scikit-learn
pandas
boto3
python-dotenv
PyGithub
flask_httpauth
dvc[s3]
catboost